move some methods to OpenTrade and change to diffent files
